Moved deduplication unit tests to its own test class
